requirements.txt
devpi-client==2.0.3
devpi-server==2.1.2
twitter.common.contextutil==0.3.2
six==1.8.0

nose
mock
coverage
